feat :  logout functionality
