+++ src/components/Logout/Logout.css
@@ -1,0 +1,9 @@
+.logoutParent {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 90vh;
+  width: 100%;
+  margin-top: 10vh;
+  /* margin-right: 20vw; */
+}
