--- src/pages/Login/Login.css
@@ -1,52 +1,0 @@
-@import "https://csscomponents.netlify.app/style.css";
-
-.loginParent {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 90vh;
-  width: 100%;
-  margin-top: 10vh;
-}
-
-.loginBody {
-  margin-left: 20%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 5rem 0;
-  width: 30rem;
-  border-radius: 1rem;
-  box-shadow: 0px 0px 14px 0px #3b82f6;
-  -webkit-box-shadow: 0px 0px 14px 0px #3b82f6;
-  -moz-box-shadow: 0px 0px 14px 0px rgba(0, 0, 0, 0.75);
-}
-
-.loginBody h1 {
-  margin-bottom: 1rem;
-  color: #3b82f6;
-}
-
-.inputBox {
-  padding: 0.5rem;
-  border-radius: 0.5rem;
-  border: 0.1rem solid #3b82f6;
-  margin: 0.5rem;
-  width: 60%;
-}
-
-input:focus-visible {
-  outline: none;
-}
-
-.btnPrimary {
-  margin-top: 4rem;
-  padding: 0.5rem;
-  border-radius: 0.3rem;
-  margin: 0.5rem 0;
-  font-size: large;
-}
-
-.errorMessage {
-  padding: 1rem;
-}
